backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
playwright>=1.40.0
aiohttp>=3.9.0
multidict>=6.0.0
attrs>=25.0.0
yarl>=1.22.0
propcache>=0.4.0
aiohappyeyeballs>=2.6.0
aiosignal>=1.4.0
frozenlist>=1.8.0
pyee>=13.0.0
greenlet>=3.2.0
litellm>=1.78.0
httpx
emergentintegrations
beautifulsoup4>=4.12.0
lxml>=5.0.0
openai
fastuuid
pyyaml
jinja2
